refactor OrderViewController to MVVM
