+++ Foodie/Foodie/Networking/NetworkService.swift
@@ -56,8 +56,6 @@ struct NetworkService {
             var result: Result<Data, Error>?
             if let data = data {
                 result = .success(data)
-                let responseString = String(data: data, encoding: .utf8)
-                print("\(responseString ?? "")")
             } else if let error = error {
                 result = .failure(error)
                 print(error.localizedDescription)
