+++ Foodie/Foodie/ViewModels/OrderViewModel.swift
@@ -1,0 +1,29 @@
+//
+//  OrderViewModel.swift
+//  Foodie
+//
+//  Created by omokagbo on 05/08/2021.
+//
+
+import Foundation
+
+class OrderViewModel {
+    
+    var orders: [Order] = []
+    
+    var notifyError: ((String) -> Void)?
+    var notifyCompletion: (() -> Void)?
+    
+    func fetchAllOrders() {
+        NetworkService.shared.fetchAllOrders { [weak self] result in
+            switch result {
+            case .success(let orders):
+                self?.orders = orders
+                self?.notifyCompletion?()
+            case .failure(let error):
+                self?.notifyError?(error.localizedDescription)
+            }
+        }
+    }
+    
+}
