+++ Foodie/Foodie/Views/Orders/OrdersViewController.swift
@@ -8,10 +8,10 @@
 import UIKit
 
 class OrdersViewController: UIViewController {
-
+    
     @IBOutlet weak var ordersTableView: UITableView!
     
-    var orders: [Order] = []
+    let viewModel = OrderViewModel()
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -22,26 +22,27 @@ class OrdersViewController: UIViewController {
     
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(true)
-        fetchAllOrders()
-    }
-    
-    private func fetchAllOrders() {
-        NetworkService.shared.fetchAllOrders { [weak self] result in
-            switch result {
-            case .success(let orders):
-                self?.dismissHUD()
-                self?.orders = orders
-                DispatchQueue.main.async {
-                    self?.ordersTableView.reloadData()
-                }
-            case .failure(let error):
-                self?.showHUDError(status: error.localizedDescription)
-            }
-        }
+        setupViewModelListeners()
     }
     
     private func registerCells() {
         ordersTableView.register(UINib(nibName: DishListTableViewCell.identifier, bundle: nil), forCellReuseIdentifier: DishListTableViewCell.identifier)
+    }
+    
+    private func setupViewModelListeners() {
+        viewModel.fetchAllOrders()
+        viewModel.notifyCompletion = { [weak self] in
+            DispatchQueue.main.async {
+                self?.dismissHUD()
+                self?.ordersTableView.reloadData()
+            }
+        }
+        
+        viewModel.notifyError = { [weak self] message in
+            DispatchQueue.main.async {
+                self?.showHUDError(status: message)
+            }
+        }
     }
 }
 
@@ -51,12 +52,12 @@ extension OrdersViewController: UITableViewDataSource {
         guard let cell = ordersTableView.dequeueReusableCell(withIdentifier: DishListTableViewCell.identifier, for: indexPath) as? DishListTableViewCell else {
             return UITableViewCell()
         }
-        cell.setup(with: orders[indexPath.row])
+        cell.setup(with: viewModel.orders[indexPath.row])
         return cell
     }
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return orders.count
+        return viewModel.orders.count
     }
     
 }
@@ -64,7 +65,7 @@ extension OrdersViewController: UITableViewDataSource {
 extension OrdersViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let controller = HomeDetailsViewController.instantiate(storyboardName: "HomeDetails")
-        controller.viewModel.dish = orders[indexPath.row].dish
+        controller.viewModel.dish = viewModel.orders[indexPath.row].dish
         navigationController?.pushViewController(controller, animated: true)
     }
 }
